Add \r to prints manually
mbed stdout->serial implementation doesn't output CR before LF, but
terminals tend to want this - they usually interpret LF as just moving the
cursor down.

Hence any mbed code printing to stdout currently has to use "\r\n" to move
the cursor to the start of the next line - this idiom is pervasive in mbed
code, but should not be necessary as \n is defined by the C standard as
moving to the start of the next line.

Register a custom trace printer that has the required extra \r hack.

Problem is usually concealed by the fact that the trace formatter normally
outputs an explicit \r and VT100 line blank at the start of each line
(presumably to overwrite ongoing command input), but it becomes apparent
if someone mixes trace and non-trace output.

Is there any chance of fixing mbed's _sys_write that underlies stdio?

On POSIX, the program would still output \n as LF, but the output device
would have ONLCR enabled to map that to CR-LF. Here, where we are writing
directly to a raw serial device, it would make sense for _sys_write to
add CR if outputting to serial in text mode. (This is what we do when
porting Nanostack to non-mbed systems).